+++ python/ray/data/_internal/pandas_block.py
@@ -16,6 +16,7 @@ from typing import (
 import numpy as np
 
 from ray.air.constants import TENSOR_COLUMN_NAME
+from ray.data._internal.row import TableRow
 from ray.data._internal.table_block import TableBlockAccessor, TableBlockBuilder
 from ray.data._internal.util import find_partitions
 from ray.data.block import (
@@ -27,7 +28,6 @@ from ray.data.block import (
     U,
 )
 from ray.data.context import DataContext
-from ray.data.row import _TableRow
 
 if TYPE_CHECKING:
     import pandas
@@ -50,7 +50,7 @@ def lazy_import_pandas():
     return _pandas
 
 
-class PandasRow(_TableRow):
+class PandasRow(TableRow):
     """
     Row of a tabular Dataset backed by a Pandas DataFrame block.
     """
